backend/ai/requirements.txt
google-generativeai
mysql-connector-python
python-dotenv
numpy
scikit-learn